Added Velocity plus more.
Made it so you have to go at a certain speed to destroy the buildings now. Also I shrunk the rubble effect a bit so that it doesn't blind the player, added a new background sky. I think a night theme would be nice for our game since we can lights and stuff down the road. We can add in a moon later.