Gas Station Explosions
I made it so the gas stations now explode when hit by the player or vehicles. The explosions also damage the buildings around it.

There is a sphere collider around the explosion prefab with a radius of 1.5 that can be changed later.

I also added a sound radius to the explosion.