Finished Setting Up Helicopters
I finished making the scripts for the helicopters. They no move to the players position and will fire missles at the player. I also made it so they spawn at the bases and then have to fly up to the right position before chasing the player. The helicopter blades also rotate and have a windup before the helicopter gets off the ground. I will add effects, sound, and interactablility next.