Finished the Tanks and Helicopters
I added all of the effects and sound to the tanks and helicopters. Also made it so the tank will rotate towards the player and angle the shot at them.

All of the projectiles also have sound and effects when destroyed.

I replaced Nathan's shooting script with my own but his is still in there just in case. I just have it deactivated.

I added a spawn limit to the number of enemies that can be in the game at all times. 7 tanks and 3 helicopters right now.

Finally, I made the damage and speed value for the tanks and helicopters projectiles public variables that can be changed in the prefab to make changing the damage values easier in the long run.