Updated Enemies and More
I believe I fixed the problem with the tanks not being abled to be picked up by turning off the nav mesh on the tanks.

I also increased the obsticle radius of the helicopters so they don't bump into each other.

I added a very basic moon and just made it a child of the player for now and added black spots to the tunnels.