Damage Indicator and Turrets
Completed the damage indicator and the turrets on the military bases.

The damage indicator will point and rotate in the direction of where the player was hit from and will fade away after about 5 seconds.

The turrets will turn and lockon to the player when within range and and stop shooting when the player leaves the range.

The turrets also have a crude difficulty setting on them for easy, normal, and hard to have no tanks when the dificulty is set to 0, one turret on dificulty 1, and two turrets at dificulty 2.