Radio Tower And Hot Dog Truck
I've fixed the tank shooting bug by setting the bullet spawner as inactive.

The radio towers are now destructible and I will add the ability to for them to drop powerups once one of them is done.

I added the hotdog truck with all the required components except for movement on the map just yet.

I also got rid of the row of cars and now only have the hotdog truck there for now and added some of the buses throughout the map.